+++ lib/vagrant/machine.rb
@@ -72,7 +72,7 @@ module Vagrant
     # @return [Vagrantfile]
     attr_reader :vagrantfile
 
-    # The SSH information for accessing this machine. 
+    # The SSH information for accessing this machine.
     # This attribute is set only when the machine is ready for SSH communication.
     #
     # @return [Hash]
@@ -391,9 +391,6 @@ module Vagrant
     #
     # @return [Hash] SSH information.
     def ssh_info
-
-      return @ssh_info unless @ssh_info.nil?
-
       # First, ask the provider for their information. If the provider
       # returns nil, then the machine is simply not ready for SSH, and
       # we return nil as well.
@@ -461,8 +458,7 @@ module Vagrant
         end
       end
 
-      # Memoize the final compiled SSH info data and return it
-      @ssh_info = info
+      info
     end
 
     # Returns the state of this machine. The state is queried from the